extract board generation out of rules.Ruleset
