+++ ruleset.go
@@ -46,14 +46,6 @@ type Snake struct {
 	EliminatedBy    string
 }
 
-type BoardState struct {
-	Height  int32
-	Width   int32
-	Food    []Point
-	Snakes  []Snake
-	Hazards []Point
-}
-
 type SnakeMove struct {
 	ID   string
 	Move string
@@ -61,7 +53,7 @@ type SnakeMove struct {
 
 type Ruleset interface {
 	Name() string
-	CreateInitialBoardState(width int32, height int32, snakeIDs []string) (*BoardState, error)
+	ModifyInitialBoardState(initialState *BoardState) (*BoardState, error)
 	CreateNextBoardState(prevState *BoardState, moves []SnakeMove) (*BoardState, error)
 	IsGameOver(state *BoardState) (bool, error)
 }
